Make container image path a dynamic value
The container image path was originally hardcoded in. We want that to be
something configurable from the SmartGateway object when it is created
in OpenShift. This change sets a default container image path that points
at the new quay.io location for the Smart Gateway and makes it something
that can be overridden.

Closes #1