Provide ability to launch an events consumer (SG)
Separate out the tasks and configuration for the two modes of Smart
Gateway; metrics and events.

Provide configuration for events and validate that both the metrics
server and the events server can be loaded simultaneously. Validate
that the pods start up.

Currently an issue with the events consumer smart gateway connecting
to the ElasticSearch server within the cluster, but I don't believe
the fix is in the smart-gateway-operator code base.

Closes #4