Bump smartgateway.infra.watch API ver. to v2alpha1 (#22)
Clean up and changes to how the Smart Gateway Operator work
in OCP4 I think deserves a version bump. The current API version
and channel for Smart Gateway Operator are related to the
Tech Preview of Service Assurance Framework as version 1.

As the OCP deployment is effectively going to SAF v2.x it makes
sense to just move everything forward. The added benefit of this
is that we don't conflict with the current channel 0.1.0 of
Smart Gateway Operator.

At some point we'll figure out how to merge the current master to
the 0.1.0 channel and this to 0.2.0, living in harmony.

Changes:
  * Drop ServiceMonitor from SGO
  * Changes to support 0.2.0 deployment
  * Fix Deployment and update containerImage values

Further work to be done, but this provides us the foundation
of the 0.2.0 SGO. Current status is good enough for development
and deployment on OCP4 with the 0.1.0 channel of Service
Assurance Operator.