Migrate to olm-catalog from gen-csv
Instead of doing a bunch of manual work and synchronizing things, leverage
the ability to use 'operator-sdk olm-catalog gen-csv' to build the CSV
from the deploy/crds/ directory and files. Doing it this will will allow
us to keep things in sync by running an --update-crds flag whenever we
adjust the example CRs for metrics and events, or the CRD for SmartGateway
in the deploy/crds/ directory.