Merge 0.2.0 back into master (#27)
* README skeleton update for OCP 4.2 development (#21)

* Bump smartgateway.infra.watch API ver. to v2alpha1 (#22)

Clean up and changes to how the Smart Gateway Operator work
in OCP4 I think deserves a version bump. The current API version
and channel for Smart Gateway Operator are related to the
Tech Preview of Service Assurance Framework as version 1.

As the OCP deployment is effectively going to SAF v2.x it makes
sense to just move everything forward. The added benefit of this
is that we don't conflict with the current channel 0.1.0 of
Smart Gateway Operator.

At some point we'll figure out how to merge the current master to
the 0.1.0 channel and this to 0.2.0, living in harmony.

Changes:
  * Drop ServiceMonitor from SGO
  * Changes to support 0.2.0 deployment
  * Fix Deployment and update containerImage values

Further work to be done, but this provides us the foundation
of the 0.2.0 SGO. Current status is good enough for development
and deployment on OCP4 with the 0.1.0 channel of Service
Assurance Operator.

* Update resources list per scorecard test

* Add operator-sdk scorecard that matches current CRD

* Align naming of CR and CRD files with operator-sdk preferrence

* Migrate to olm-catalog from gen-csv

Instead of doing a bunch of manual work and synchronizing things, leverage
the ability to use 'operator-sdk olm-catalog gen-csv' to build the CSV
from the deploy/crds/ directory and files. Doing it this will will allow
us to keep things in sync by running an --update-crds flag whenever we
adjust the example CRs for metrics and events, or the CRD for SmartGateway
in the deploy/crds/ directory.

* Update README for latest changes

* Fully pass operator-courier verify --ui_validate_io

* Initial commit

* Update source code URL

* Update package name to not conflict with container registry

* Update CSV to match gen-csv output

* Add basic script for updating the CSV

* Script for pushing changes to application registry

* Enable UseTimestamp true by default on metrics SG

Enable timestamps by default when setting up a metrics Smart Gateway

* Address some documentation <todo> items (#28)