Complete path to container image
Previous commit I missed 4 locations that referenced the container image path. I leveraged
grep to find all of the instances and so we should be good now. I adjusted deploy/operator.yaml
manually and then ran:

  operator-sdk olm-catalog gen-csv --csv-version 0.2.0 --update-crds

Version of operator-sdk was v0.12.0. Executing that command resulted in the container
image paths being updated in the olm-catalog/