use_timestamp is a string from smart gateway viewpoint
Setting the CRD validation to boolean instead of string causes a mismatch between the expected
values that the CRD validation is expecting, and what the Smart Gateway is expecting. The smart
gateway expects the values to be passed to it as a string (looks for literal true/false vs
a boolean true/false).

This change makes the CRD validation reflect what the Smart Gateway expects to be passed.

Closes https://github.com/infrawatch/service-telemetry-operator/issues/65