Fix broken type validation in CRD
Fix broken type validation in the CRD. The Smart Gateway expects some boolean
looking values to actually be passed as a string. Some recent changes in the CRD
to clean things up has resulted in setting the validation types to boolean when
the SG actually expects them to be passed as literal strings.

This change updates the CRD and executes the ./build/update_csv.sh script to
keep everything in sync.