Migrate to new AMQP1Connections configuration options
Migrate to using the new AMQP1Connections configuration options which allows for the
setup of the URL and DataSource. The use of DataSource is for when we're running multiple
Smart Gateway instances for different data sources, which allows for the creation of
indexes and such that matches the expected data source.

Closes #42