Migrate to new AMQP1Connections configuration options (#43)
* Migrate to new AMQP1Connections configuration options

Migrate to using the new AMQP1Connections configuration options which allows for the
setup of the URL and DataSource. The use of DataSource is for when we're running multiple
Smart Gateway instances for different data sources, which allows for the creation of
indexes and such that matches the expected data source.

Closes #42

* Fix syntax for array of AMQP1Connections

Misinterpreted the struct that it was an array of objects. Updated to reflect that
along with fixing an outstanding indenting bug which made the ConfigMap a bit
ugly (although it was perfectly functional).

* Revert metrics ConfigMap back to AMQPMetricURL