Migrate to using Deployments
Migrate to using Deployment instead of DeploymentConfig to make things better for
vanilla k8s instead of using OpenShift specific extensions. Should provide us a
path to using molecule for functional tests in the future.

Also adjusts the method used to load the templates to a more modern interface
that is cleaner and easier to maintain and extend.