Set Smart Gateway container image from env var
Setting the container image for the Smart Gateway from an environment variable allows
for setting a specific version of the Smart Gateway to come up, allowing versions to
be tied to the Operator container image itself, and thus a CSV version.

Looking at other Operators, any dependent images seem to get set via environment variable
on the Operator, and then used for instantiating the desired set of images for that CSV.

If the SMARTGATEWAY_IMAGE environment variable is not set, then the default is to use
the latest image from quay.io/infrawatch/smart-gateway.