Reload Smart Gateways when ConfigMap changes (#64)
* Reload Smart Gateways when ConfigMap changes

Changes to support ability to reload Smart Gateways when the ConfigMap changes. Uses
the k8s_config_resource_name module along with the append_hash option of the k8s module
to provide a unique name for the ConfigMap so that it can be referenced by the Deployment.

When the contents of the ConfigMap change, the hash changes, resulting in a new ConfigMap
being created, and a new resource name being used in the Deployment. When the Deployment is
changed, the signals Kubernetes to update the Deployment, resulting in a new pod running
which results in the configuration file being parsed again.

Still needs some additional logic because this current iteration results in multiple
ConfigMaps being created and old ones not cleaned up.

* Update label vs ConfigMap name

Use the same technique that was applied to a ConfigMap rename, and instead just set
the resource config name against a label in the Deployment. Thanks to Chris Sibbitt
for the suggestion.