Create CSV directory structure to pass validations
The directory structure is really picky when you plan to do a bunch of validations
and still maintain a multi-CSV version structure. There are a few duplicates of things
but all that data is generated from the root of each CSV version.

Creating the directory structure and files in this way allows building of bundles,
validation of the bundle directories directly, creation of the index image for
the CatalogSource, and passing operator-courier verify.

One workaround is required, that results in the smart-gateway-operator.package.yaml
being moved out of the deploy/olm-catalog/smart-gateway-operator/ directory prior
to building and generating, or else things get confused.