Migrate latest CSV to manifests directory
Migrate latest CSV to manifests directory in preparation for migrating to the
bundle release format vs the application registry approach. Update the operator.yaml
which was out of sync with the latest CSV.

Latest CSV has been updated to sync with the operator.yaml which has the current
container image tags applied to the CSV which was incorrect.