bundle migration (#67)
* Migrate latest CSV to manifests directory

Migrate latest CSV to manifests directory in preparation for migrating to the
bundle release format vs the application registry approach. Update the operator.yaml
which was out of sync with the latest CSV.

Latest CSV has been updated to sync with the operator.yaml which has the current
container image tags applied to the CSV which was incorrect.

* Create initial bundle

* Update Travis-CI to check bundle validation

* No automatic releases for right now

* Add new README.md for release process

* Be verbose when checking bundle validation

* Update build/README.md for validation command

* Smart Gateway Operator v2.0.1 is the base release

Remove the replaces line from the CSV to start our base release for bundles as
the most current release. From here we will create new CSVs that will replace
starting at v2.0.1 rather than having to import all the historical CSVs which
are not used by anyone.