Smart Gateway Operator API v2 interface (#84)
* Smart Gateway Operator API v2 interface

Update the Smart Gateway Operator API (CRD) interface to v2, which mimics the re-written
sg-core application. Additional plugins and configuration are now available and more will
be written over a period of time. Changing the interface provides more flexibility in the
future and allows others who add plugins to the sg-core application to continue making use
of the Smart Gateway Operator.

Resolves: #81
Closes: rhbz#1959594

* Update README.md

* Just use Service name passed

* Fix ports in deployment

* Fix port defines in Deployment for services

* Add resources back to CSV for informational purposes