Migrate to operator-sdk 1.16.0
Migrate the Smart Gateway Operator from operator-sdk 0.19.4 to
operator-sdk 1.16.0 for OpenShift v4.9 and later.

A lot of changes here are boilerplate as generated by operator-sdk init
and operator-sdk create api commands. Once the boilerplate was created,
then the existing Ansible role was moved over. As the Ansible Operator
framework is now using collections, the name of the k8s module and
filters were updated to reflect this in the Ansible roles/ directory.

RBAC changes were also implemented to satisfy for the existing RBAC
rules that were in place prior to the migration. A new smart-gateway
role and service account were created in order to satisfy for the
running sg-core workload along with the lated oauth-proxy image shipping
with OCP v4.9.

GitHub actions have been updated to reflect the changes to the
./build/generate_bundle.sh script, which was updated to use the new
Makefile approach to building and running the Operator.

Additionally, the sample CustomResource for a SmartGateway has been
changed to leverage the dummy metrics plugin removing the requirement to
have a running AMQ Interconnect router system and data source. This
should make testing via molecule (or otherwise) easier in the future,
allowing for component testing outside of system e2e testing.