Test via molecule with KinD
Implement a basic molecule test with KinD in order to load the operator
and the workload, then perform a basic assertion test to verify that
everything is operational. More work will be required here as the test
is effectively a stub at this point.

In the molecule test needed to fake the creation of a TLS certificate
that is normally dealt with in OpenShift without this effort. Created a
locally self-signed certificate that is then loaded as the
smartgateway-sample-proxy-tls Secret that is mounted by the Smart
Gateway Deployment as a mounted volume for oauth-proxy.

In an OpenShift environment the oauth-proxy is available in the local
registry. Within a KinD environment that is not available. The existing
oauth_proxy_image path is defaulted in the role, but is now exposed as a
parameter to the SmartGateway custom resource. Because of the migration
to CRD/v1 unknown parameters are no longer preserved by default. This
new parameter is exposed primarily for testing but should be useful in
the future as well.

Adds another Makefile option called kustomize-local which doesn't have
the same checks as the standard kustomize command. In the testing
environment an incompatible version of kustomize may be installed, and
it's not always possible to remove it. Using kustomize-local results in
a specific kustomize version being added locally and then referenced by
the testing infrastructure for consistency.