Migrate to operator-sdk 1.16.0 (#108)
* Migrate to operator-sdk 1.16.0

Migrate the Smart Gateway Operator from operator-sdk 0.19.4 to
operator-sdk 1.16.0 for OpenShift v4.9 and later.

A lot of changes here are boilerplate as generated by operator-sdk init
and operator-sdk create api commands. Once the boilerplate was created,
then the existing Ansible role was moved over. As the Ansible Operator
framework is now using collections, the name of the k8s module and
filters were updated to reflect this in the Ansible roles/ directory.

RBAC changes were also implemented to satisfy for the existing RBAC
rules that were in place prior to the migration. A new smart-gateway
role and service account were created in order to satisfy for the
running sg-core workload along with the lated oauth-proxy image shipping
with OCP v4.9.

GitHub actions have been updated to reflect the changes to the
./build/generate_bundle.sh script, which was updated to use the new
Makefile approach to building and running the Operator.

Additionally, the sample CustomResource for a SmartGateway has been
changed to leverage the dummy metrics plugin removing the requirement to
have a running AMQ Interconnect router system and data source. This
should make testing via molecule (or otherwise) easier in the future,
allowing for component testing outside of system e2e testing.

* Bump maxOpenShiftVersion to 4.10

Bump the maxOpenShiftVersion to OCP 4.10 which is the target for the next y-stream release of STF.

* Resolve issues with CI testing

* Ignore bin/ directory holding local kustomize binary

* Resolve indentation inconsistency

* Reduce RBAC permission for Smart Gateway instances

* Drop superfluous resource from authentication API group

* Bump operator version annotation

* Test via molecule with KinD (#109)

* Test via molecule with KinD

Implement a basic molecule test with KinD in order to load the operator
and the workload, then perform a basic assertion test to verify that
everything is operational. More work will be required here as the test
is effectively a stub at this point.

In the molecule test needed to fake the creation of a TLS certificate
that is normally dealt with in OpenShift without this effort. Created a
locally self-signed certificate that is then loaded as the
smartgateway-sample-proxy-tls Secret that is mounted by the Smart
Gateway Deployment as a mounted volume for oauth-proxy.

In an OpenShift environment the oauth-proxy is available in the local
registry. Within a KinD environment that is not available. The existing
oauth_proxy_image path is defaulted in the role, but is now exposed as a
parameter to the SmartGateway custom resource. Because of the migration
to CRD/v1 unknown parameters are no longer preserved by default. This
new parameter is exposed primarily for testing but should be useful in
the future as well.

Adds another Makefile option called kustomize-local which doesn't have
the same checks as the standard kustomize command. In the testing
environment an incompatible version of kustomize may be installed, and
it's not always possible to remove it. Using kustomize-local results in
a specific kustomize version being added locally and then referenced by
the testing infrastructure for consistency.

* Remove extra --namespace parameter from molecule test

* Update documentation and metadata

Update the README.md to match the currently known methods to install
Smart Gateway Operator using the new operator-sdk migration. Update the
metadata to cover the new version of the Operator from 4.0 to 5.0 major
release number.

* Update LICENSE copyright date range

* Adjust documentation and bundle script generator