Revert "Migrate to operator-sdk 1.16.0 (#108)"
Revert changes to migrate Smart Gateway Operator to operator-sdk 1.16.0
until the same set of changes can be made to Service Telemetry Operator.

Changing how the operator is built causes a regression in the testing
framework for STF. Instead of making temporary changes to the testing
framework to deal with the changes in Operator-SDK 1.16.0, we'll make
changes to the testing framework when migrating Service Telemetry
Operator to operator-sdk 1.16.0 as well.

In the meantime, revert these changes which will just end up landing in
conjunction with the changes to Service Telemetry Operator as a bundled
change set.

This reverts commit 8328ed835db1afd9d33120cc19badf7773f4d86c.