Add requires infrastructure annotations (#146)
* Add requires infrastructure annotations

Add required infrastructure annotations for the bundle. Implementation
is done in generate_bundle.sh because the annotations.yaml file in the
deploy/olm-catalog/ directory is not read by operator-sdk-0.19.4. Append
required additional feature annotations to the generated
annotations.yaml by operator-sdk generate bundle.

Related STF-1530

* Revert "Add requires infrastructure annotations"

This reverts commit 8a9b2c66a81a753301adae383fc24963a7ed9c9c.

* Add required infrastructure annotations

Add the required infrastructure annotations to the CSV directly, as that
is the expected location for the new annotations.

Related: STF-1530

* Generate CSV contents with operator-sdk